Add support for the old() keyword.
This works by assigning values to the current context before the function is called and substituting references to them in the rules. Each old reference is reffered to by the executing id of the current contract check to prevent collisions in self-referential functions.

Also note, this has required the usage of underscore.js

References issue #1